Move app to separate file, handle cli.ExitCoder errors
